move plotting -> scoring
