+++ tests/unit_tests/search/scoring/test_plotting.py
@@ -1,8 +1,8 @@
 import matplotlib
 import numpy as np
 
-from alphadia.plotting.cycle import plot_cycle
 from alphadia.plotting.utils import lighten_color
+from alphadia.search.scoring.plotting.cycle import plot_cycle
 
 
 def test_lighten_color():
